+++ packages/types/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Redundant APIWrapper interface (#124)
+
+### Changed
+- Global injected apis (#124)
 
 ## [4.4.0] - 2025-04-24
 ### Changed